feat: advertize nodeport services via mdns
k8s NodePort Services that are annotated with `kubemate.mgoltzsche.github.com/mdns-service` are advertized within the host network/LAN via mDNS.
This is to make e.g. the Spotify and Snapcast mobile apps find the corresponding service (deployed as kubemate app) within the LAN/wifi network.

As mDNS implementation brutella/dnssd is used instead of hashicorp/mdns because the latter is not compatible with zeroconf/avahi/bonjour as opposed to the former.